bump to edition 2024

+++ src/vpi.rs
@@ -4,7 +4,7 @@ use std::ptr;
 /// Get current simulation time in _simulation time unit_
 pub fn get_time() -> u64 {
     let mut time = sys::s_vpi_time {
-        type_: sys::vpiSimTime as i32,
+        type_: sys::vpiSimTime,
         high: 0,
         low: 0,
         real: 0.0,
